// Technical-Department-front/src/app/kitchen/model/day-of-week-tab.ts
export interface DayOfWeekTab{
    id?: number;
    name: string;
    date: Date;
    breakfastChecked: boolean;
    lunchChecked: boolean;
    dinnerChecked: boolean;
    morningSnackChecked: boolean;
    dinnerSnackChecked: boolean;
    lunchSaladChecked: boolean;
    dinnerSaladChecked: boolean;
}